+++ src/components/mypage/EditProfile.tsx
@@ -2,18 +2,18 @@
 
 import { checkUserNickname, getUserInfo, updateUserInfo } from '@/app/api/mypage/user-api';
 import { useLoginStore } from '@/store/login/loginUserIdStore';
-import { userInfoStore } from '@/store/userInfoStore';
-import { UpdateUserInfoType } from '@/types/user';
-import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { UpdateUserInfoType, UserInfoType } from '@/types/user';
+import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 import React, { useEffect, useState } from 'react';
 import { ToastContainer } from 'react-toastify';
 import { noChangedNotify } from '../common/Toastify';
 import EditProfileImage from './EditProfileImage';
-import { supabase } from '@/app/api/supabase/supabase';
+import { userInfoStore } from '@/store/userInfoStore';
 
 const EditProfile = () => {
   const { loginUserId } = useLoginStore();
   const userId = loginUserId as string;
+  // console.log('userId', userId);
 
   const queryClient = useQueryClient();
 
@@ -21,30 +21,30 @@ const EditProfile = () => {
   const { userInfo, setUserInfo } = userInfoStore();
 
   useEffect(() => {
-    const fetchUserInfo = async () => {
-      if (userId) {
+    if (userId) {
+      const fetchUserInfo = async () => {
         const userInfoData = await getUserInfo({ userId });
         if (userInfoData !== null) {
           setUserInfo(userInfoData);
         }
-      }
-    };
-    fetchUserInfo();
+      };
+      fetchUserInfo();
+    }
   }, [userId, setUserInfo]);
+  // console.log('userInfo', userInfo);
 
   const [newNickname, setNewNickname] = useState('');
   const [newProfileImage, setNewProfileImage] = useState('');
   const [isEditing, setIsEditing] = useState(false); // 수정된 사항 확인 여부
   const [isAvailableNickname, setIsAvailableNickname] = useState(true); // 닉네임 중복 여부 상태 업데이트
   const [isActiveBtn, setIsActiveBtn] = useState(false); // 수정 완료시 버튼 활성화 상태
-  const [selectedImage, setSelectedImage] = useState<File>(null!); // 선택된 이미지 파일
 
   useEffect(() => {
     if (userInfo) {
       setNewNickname(userInfo.nickname || '');
       setNewProfileImage(userInfo.profile_image || '');
     }
-  }, [userInfo, setNewNickname, setNewProfileImage]);
+  }, [userInfo]);
 
   // 닉네임 수정
   const handleOnChangeNickname = async (e: React.ChangeEvent<HTMLInputElement>) => {
@@ -70,39 +70,20 @@ const EditProfile = () => {
     }
   });
 
-  // supabase storage에 프로필 이미지 업로드
-  const uploadProfileImage = async (file: File) => {
-    const randomUUID = crypto.randomUUID();
-    const filePath = `profile/${randomUUID}`;
-    const { data, error } = await supabase.storage.from('profileImages').upload(filePath, file);
-    if (error) {
-      console.error('파일 업로드 실패 :', error);
-      throw error;
-    } else {
-      const url = `${process.env.NEXT_PUBLIC_SUPABASE_URL}/storage/v1/object/public/profileImages/${data.path}`;
-      // setNewProfileImage(url);
-      return url;
-    }
-  };
-
   // 수정하기 버튼 -> supabase에 수정한 정보 update
-  const handleOnClickEditProfileBtn = async () => {
-    const uploadImageUrl = await uploadProfileImage(selectedImage);
-    setNewProfileImage(uploadImageUrl);
-
+  const handleOnClickEditProfileBtn = () => {
     // 수정된 사항이 없는 경우
     const isNicknameChanged = newNickname !== userInfo?.nickname;
-    const isProfileImageChanged = newProfileImage !== userInfo?.profile_image;
+    const isProfileImageChanged = newProfileImage != userInfo?.profile_image;
 
     if (!isNicknameChanged && !isProfileImageChanged) {
       noChangedNotify();
       return;
-    } else {
-      // 수정된 사항이 있는 경우 : supabase table에 update
-      updateUserInfoMutation({ newNickname, newProfileImage });
-      alert('프로필 수정이 완료되었습니다.');
-      setIsEditing(false);
     }
+
+    // 수정된 사항이 있는 경우
+    updateUserInfoMutation({ newNickname, newProfileImage });
+    alert('프로필 수정이 완료되었습니다.');
   };
 
   // 취소하기 버튼
@@ -119,6 +100,10 @@ const EditProfile = () => {
     }
   };
 
+  // if (isPending) {
+  //   return <div> 로딩중 ... </div>;
+  // }
+
   if (!userInfo) {
     return <div> 유저 정보가 없습니다.</div>;
   }
@@ -127,10 +112,8 @@ const EditProfile = () => {
     <div className="flex">
       <EditProfileImage
         newProfileImage={newProfileImage}
+        setNewProfileImage={setNewProfileImage}
         isEditing={isEditing}
-        selectedImage={selectedImage}
-        setSelectedImage={setSelectedImage}
-        userInfo={userInfo}
       />
       <div className="flex flex-col">
         <div className="flex flex-col">
