+++ src/components/mypage/EditProfileImage.tsx
@@ -1,25 +1,15 @@
 import { supabase } from '@/app/api/supabase/supabase';
 import Image from 'next/image';
-import { ChangeEvent, Dispatch, SetStateAction, useRef, useState } from 'react';
+import { ChangeEvent, Dispatch, SetStateAction, useRef } from 'react';
 import basicProfileImage from '../../../public/profile-image.png';
-import { useStartTyping } from 'react-use';
-import { UserInfoType } from '@/types/user';
 
 interface EditProfileImageProps {
   newProfileImage: string;
+  setNewProfileImage: Dispatch<SetStateAction<string>>;
   isEditing: boolean;
-  selectedImage: File | null;
-  setSelectedImage: React.Dispatch<React.SetStateAction<File | null>>;
-  userInfo: UserInfoType;
 }
 
-const EditProfileImage = ({
-  newProfileImage,
-  isEditing,
-  selectedImage,
-  setSelectedImage,
-  userInfo
-}: EditProfileImageProps) => {
+const EditProfileImage = ({ newProfileImage, setNewProfileImage, isEditing }: EditProfileImageProps) => {
   const fileInput = useRef<HTMLInputElement | null>(null);
 
   // 프로필 이미지 수정 버튼 클릭
@@ -27,21 +17,33 @@ const EditProfileImage = ({
     fileInput.current?.click();
   };
 
-  // 수정된 프로필 이미지 반영
-  const handleOnChangeImage = (e: ChangeEvent<HTMLInputElement>) => {
-    const file: File | undefined = e.target.files?.[0];
-    // console.log('file', file);
-    if (file) {
-      setSelectedImage(file);
+  // supabase storage에 프로필 이미지 업로드
+  const uploadProfileImage = async (file: File) => {
+    const randomUUID = crypto.randomUUID();
+    const filePath = `profile/${randomUUID}`;
+    const { data, error } = await supabase.storage.from('profileImages').upload(filePath, file);
+    if (error) {
+      console.error('파일 업로드 실패 :', error);
+      throw error;
     } else {
-      setSelectedImage(null);
+      const url = `${process.env.NEXT_PUBLIC_SUPABASE_URL}/storage/v1/object/public/profileImages/${data.path}`;
+      // console.log('url', url);
+      return setNewProfileImage(url);
     }
   };
 
+  // 수정된 프로필 이미지 반영
+  const handleOnChangeImage = (e: ChangeEvent<HTMLInputElement>) => {
+    const file: File | undefined = e.target.files?.[0];
+    if (!file) {
+      return;
+    }
+
+    uploadProfileImage(file);
+  };
+
   // 프로필 이미지가 없을 때, 기본 프로필 이미지 보여주기
-  const profileImage = selectedImage
-    ? URL.createObjectURL(selectedImage)
-    : newProfileImage || userInfo.profile_image || basicProfileImage;
+  const profileImage = newProfileImage ? newProfileImage : basicProfileImage;
 
   return (
     <div>
@@ -52,7 +54,7 @@ const EditProfileImage = ({
           height={100}
           className="rounded-full"
           alt="프로필 이미지"
-          unoptimized={true}
+          unoptimized={true} // 추후 수정 해야함
         />
         <input
           type="file"
